permission system added

+++ Controller/LinesController.php
@@ -79,6 +79,8 @@ class LinesController extends AppController {
     function admin_add() {
         if (!empty($this->data)) {
             $this->Line->create();
+            $this->request->data['Line']['user_id'] = $this->Auth->user('id');
+            $this->request->data['Line']['author'] = $this->Auth->user('username');
             if ($this->Line->save($this->data)) {
                 $this->Session->setFlash(__('The data has been saved', true));
                 $this->redirect(array('action' => 'index'));
@@ -89,30 +91,39 @@ class LinesController extends AppController {
     }
 
     function admin_edit($id = null) {
-	    if (!$id && empty($this->data)) {
-		$this->Session->setFlash(__('Please do following links in the page', true));
-		$this->redirect($this->referer());
-	    }
-	    if (!empty($this->data)) {
-		if ($this->Line->save($this->data)) {
-		    $this->Session->setFlash(__('The data has been saved', true));
-		    $this->redirect(array('action' => 'index'));
-		} else {
-		    $this->Session->setFlash(__('Something was wrong during saving, please try again', true));
-		}
-	    }
-	    $this->set('id', $id);
-	    $this->data = $this->Line->read(null, $id);
-     }
+        if($id == $this->Auth->user('id') || $this->Auth->user('id') == 1) {
+            if (!$id && empty($this->data)) {
+                $this->Session->setFlash(__('Please do following links in the page', true));
+                $this->redirect($this->referer());
+            }
+            if (!empty($this->data)) {
+                if ($this->Line->save($this->data)) {
+                        $this->Session->setFlash(__('The data has been saved', true));
+                        $this->redirect(array('action' => 'index'));
+                } else {
+                        $this->Session->setFlash(__('Something was wrong during saving, please try again', true));
+                }
+            }
+            $this->set('id', $id);
+            $this->data = $this->Line->read(null, $id);
+        } else {
+            $this->Session->setFlash(__('You\'re not the owner of this line. Permission Denied.', true));
+            $this->redirect(array('action' => 'index'));
+        }
+    }
 
     function admin_delete($id = null) {
+        if($id == $this->Auth->user('id') || $this->Auth->user('id') == 1) {
             if (!$id) {
                 $this->Session->setFlash(__('Please do following links in the page', true));
             } else if ($this->Line->delete($id)) {
                 $this->Session->setFlash(__('The data has been deleted', true));
             }
             $this->redirect(array('action' => 'index'));
-	
+        } else {
+            $this->Session->setFlash(__('You\'re not the owner of this line. Permission Denied.', true));
+            $this->redirect(array('action' => 'index'));
+        }
     }
 
 }
